typeck: move method errors/suggestions to their own file.
